requirements.txt
bleach==6.3.0
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
